refa(cargo-shuttle): nits, util file
